# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi application",
        "version": "1.0"
    },
    "paths": {
...
        "/api/resources": {
            "get": {
                "tags": [
                    "Resource"
                ],
                "summary": "Get all resources",
                "operationId": "getAllResources",
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Resource"
                ],
                "summary": "Post Resource",
                "operationId": "postResource",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "name": "Oil",
                                "place": "Kazan"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Everything is fine"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/resources/{resource_id}": {
            "get": {
                "tags": [
                    "Resource"
                ],
                "summary": "Get Resource by id",
                "operationId": "getResource",
                "parameters": [
                    {
                        "name": "resource_id",
                        "in": "path",
                        "description": "Resource ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Resource"
                ],
                "summary": "Update Resource by id",
                "operationId": "updateResource",
                "parameters": [
                    {
                        "name": "resource_id",
                        "in": "path",
                        "description": "Resource ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "name": "Oil",
                                "place": "Kazan"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Resource"
                ],
                "summary": "Delete Resource by id",
                "operationId": "deleteResource",
                "parameters": [
                    {
                        "name": "resource_id",
                        "in": "path",
                        "description": "Resource ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Resource"
                ],
                "summary": "Change Resource by id",
                "operationId": "changeResource",
                "parameters": [
                    {
                        "name": "resource_id",
                        "in": "path",
                        "description": "Resource ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "name": "Oil",
                                "place": "Kazan"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    }
}